Move `RunConfig` and friends to `types.ts`.
